Fix clash in emailProcessorContact
Fix two instances where the desired behaviour is the default behaviour, but we were setting $result to enforce that explicitly. As $result is passed by reference, this caused a possible clash where another implementation of the hook also wanted to return a result- the last implementation of the hook to be run was 'winning'. Also removed some cruft checking for situations that should never happen.